lessons-m3/les_21_DRF_ViewSet/habits/api/views.py
from rest_framework import mixins, permissions, viewsets
from rest_framework.permissions import IsAuthenticated, IsAuthenticatedOrReadOnly, AllowAny

from habits.models import Habit
from .serializers import HabitSerializer


class HabitViewSet(viewsets.ModelViewSet):
    permission_classes = (IsAuthenticatedOrReadOnly,)
    queryset = Habit.objects.all()
    serializer_class = HabitSerializer

    def get_queryset(self):
        return super().get_queryset().filter(user=self.request.user)

    def perform_create(self, serializer):
        serializer.save(user=self.request.user)
